Rename const db in variables.ts, move reset and verify token methods to there own controller file.
